+++ components/scheduler/src/main/java/org/cardanofoundation/ledgersync/scheduler/service/offchain/drepregistration/DRepRegistrationPersistServiceImpl.java
@@ -1,16 +1,16 @@
-package org.cardanofoundation.ledgersync.scheduler.service.offchain.committee_deregistration;
+package org.cardanofoundation.ledgersync.scheduler.service.offchain.drepregistration;
 
 import java.sql.Timestamp;
 import java.util.List;
 
-import org.cardanofoundation.ledgersync.consumercommon.entity.OffChainCommitteeDeregistration;
+import org.cardanofoundation.ledgersync.consumercommon.entity.OffChainDRepRegistration;
 import org.cardanofoundation.ledgersync.consumercommon.entity.OffChainDataCheckpoint;
 import org.cardanofoundation.ledgersync.consumercommon.entity.OffChainFetchError;
 import org.cardanofoundation.ledgersync.consumercommon.enumeration.OffChainCheckpointType;
-import org.cardanofoundation.ledgersync.scheduler.dto.anchor.CommitteeDeregistrationDTO;
+import org.cardanofoundation.ledgersync.scheduler.dto.anchor.DRepRegistrationDTO;
 import org.cardanofoundation.ledgersync.scheduler.service.offchain.OffChainProcessPersistDataService;
 import org.cardanofoundation.ledgersync.scheduler.storage.EraRepo;
-import org.cardanofoundation.ledgersync.scheduler.storage.governance.CommitteeDeregistrationRepo;
+import org.cardanofoundation.ledgersync.scheduler.storage.governance.DRepRegistrationRepo;
 import org.cardanofoundation.ledgersync.scheduler.storage.offchain.OffChainDataCheckpointStorage;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
@@ -24,14 +24,14 @@ import lombok.extern.slf4j.Slf4j;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Slf4j
 @RequiredArgsConstructor
-@Qualifier("committeeDeregPersistServiceImpl")
-public class CommitteeDeregPersistServiceImpl implements OffChainProcessPersistDataService {
+@Qualifier("dRepRegistrationPersistServiceImpl")
+public class DRepRegistrationPersistServiceImpl implements OffChainProcessPersistDataService {
 
     static final long MAX_TIME_QUERY = 432000L;
     final OffChainDataCheckpointStorage offChainDataCheckpointStorage;
-    final CommitteeDeregistrationRepo committeeDeregistrationRepo;
-    final CommitteeDeregExtractFetchService committeeDeregExtractFetchService;
-    final CommitteeDeregStoringService committeeDeregStoringService;
+    final DRepRegistrationRepo dRepRegistrationRepo;
+    final DRepRegistrationExtractFetchService dRepRegistrationExtractFetchService;
+    final DRepRegistrationStoringService dRepRegistrationStoringService;
     final EraRepo eraRepo;
 
     @Override
@@ -39,30 +39,32 @@ public class CommitteeDeregPersistServiceImpl implements OffChainProcessPersistD
         long startTime = System.currentTimeMillis();
 
         OffChainDataCheckpoint currentCheckpoint = getCurrentCheckpoint(offChainDataCheckpointStorage, eraRepo,
-                OffChainCheckpointType.COMMITTEE_DEREGISTRATION);
-        long currentSlotNo = committeeDeregistrationRepo.maxSlotNo().orElse(currentCheckpoint.getSlotNo());
+                OffChainCheckpointType.DREP_REGISTRATION);
+
+        long currentSlotNo = dRepRegistrationRepo.maxSlotNo().orElse(currentCheckpoint.getSlotNo());
         if (currentSlotNo - currentCheckpoint.getSlotNo() > MAX_TIME_QUERY) {
             currentSlotNo = currentCheckpoint.getSlotNo() + MAX_TIME_QUERY;
         }
 
-        List<CommitteeDeregistrationDTO> committeeDeregList = committeeDeregistrationRepo.getAnchorInfoBySlotRange(
+        List<DRepRegistrationDTO> dRepRegistrationList = dRepRegistrationRepo.getAnchorInfoBySlotRange(
                 currentCheckpoint.getSlotNo(), currentSlotNo);
 
-        committeeDeregExtractFetchService.initOffChainListData();
-        committeeDeregExtractFetchService.crawlOffChainAnchors(committeeDeregList);
+        dRepRegistrationExtractFetchService.initOffChainListData();
+        dRepRegistrationExtractFetchService.crawlOffChainAnchors(dRepRegistrationList);
 
-        List<OffChainFetchError> offChainFetchErrors = committeeDeregExtractFetchService.getOffChainAnchorsFetchError();
-        List<OffChainCommitteeDeregistration> offChainDataList = committeeDeregExtractFetchService
+        List<OffChainFetchError> offChainFetchErrors = dRepRegistrationExtractFetchService
+                .getOffChainAnchorsFetchError();
+        List<OffChainDRepRegistration> offChainDataList = dRepRegistrationExtractFetchService
                 .getOffChainAnchorsFetch();
 
-        committeeDeregStoringService.insertFetchData(offChainDataList);
-        committeeDeregStoringService.insertFetchFailData(offChainFetchErrors);
+        dRepRegistrationStoringService.insertFetchData(offChainDataList);
+        dRepRegistrationStoringService.insertFetchFailData(offChainFetchErrors);
 
         currentCheckpoint.setSlotNo(currentSlotNo);
         currentCheckpoint.setUpdateTime(new Timestamp(System.currentTimeMillis()));
         offChainDataCheckpointStorage.save(currentCheckpoint);
 
-        log.info("End fetching Committee Deregistration metadata, taken time: {} ms",
+        log.info("End fetching DRep Registration metadata, taken time: {} ms",
                 System.currentTimeMillis() - startTime);
     }
