+++ components/scheduler/src/main/java/org/cardanofoundation/ledgersync/scheduler/service/offchain/votingdata/VotingDataExtractFetchService.java
@@ -1,4 +1,4 @@
-package org.cardanofoundation.ledgersync.scheduler.service.offchain.voting_data;
+package org.cardanofoundation.ledgersync.scheduler.service.offchain.votingdata;
 
 import org.cardanofoundation.ledgersync.consumercommon.entity.OffChainFetchError;
 import org.cardanofoundation.ledgersync.consumercommon.entity.OffChainVotingData;
@@ -9,7 +9,7 @@ import org.cardanofoundation.ledgersync.scheduler.SchedulerProperties;
 import org.cardanofoundation.ledgersync.scheduler.dto.anchor.VotingDataAnchorDTO;
 import org.cardanofoundation.ledgersync.scheduler.dto.offchain.OffChainFetchResultDTO;
 import org.cardanofoundation.ledgersync.scheduler.dto.offchain.OffChainVotingFetchResultDTO;
-import org.cardanofoundation.ledgersync.scheduler.service.offchain.OffChainFetchAbstractService;
+import org.cardanofoundation.ledgersync.scheduler.service.offchain.OffChainFetchService;
 import org.springframework.stereotype.Component;
 
 import lombok.AccessLevel;
@@ -22,7 +22,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @RequiredArgsConstructor
 public class VotingDataExtractFetchService extends
-        OffChainFetchAbstractService<OffChainVotingData, OffChainFetchError, OffChainVotingFetchResultDTO, VotingDataAnchorDTO> {
+        OffChainFetchService<OffChainVotingData, OffChainFetchError, OffChainVotingFetchResultDTO, VotingDataAnchorDTO> {
 
     final SchedulerProperties properties;
